:truck: Separate Context from ContextSettings
