+++ data_safe_haven/context/__init__.py
@@ -1,5 +1,6 @@
+from .context import Context
 from .context_infra import ContextInfra
-from .context_settings import Context, ContextSettings
+from .context_settings import ContextSettings
 
 __all__ = [
     "Context",
